This is the 3rd version of the project and Add LectureList
Editing for verification.
Add PUT to remove student. DELETE method does not allow add extra field

If you look at the commit, I redid the project 3 times. Old project at
https://github.com/serg-lubeshko/Lubeshko_t_4.git.
The problem there was that when deleting the migration and the database,
there was an error with the creation of a new database.

I move all  models to a new folder